ai-engine/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
gitpython>=3.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
pydantic>=1.8.0
python-hcl2>=3.0.0
streamlit>=1.0.0
